backend/requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
pyzbar
pillow
google-generativeai
python-dotenv
